Jungle tree refactoring.
